0063: 7F49 to jeden dwubajtowy tag - RSA nie dawalo sie odczytac
Wystawienie na RSA padalo z "The generated key response has no tag 81" dla
klucza lezacego w odpowiedzi dokladnie tam, gdzie powinien.

Parser BER czytal wylacznie jednobajtowe tagi. W BER tag, ktorego dolne piec
bitow to same jedynki, ciagnie sie dalej - 7F 49 jest wiec jednym tagiem, a
nie tagiem 7F z dlugoscia 0x49.

Dla ECC dzialalo przypadkiem: 0x49 przekraczalo bufor, kod przycinal wartosc
do konca danych, a [1..] u wolajacego zjadalo dokladnie ten jeden bajt, ktory
trzeba bylo zjesc. Odpowiedz RSA jest dluzsza, nic nie przekracza bufora, i
parser wchodzil w srodek modulu.

Nie wyszlo to wczesniej, bo agent mial EccP256 wpisane na sztywno, wiec jedyna
odpowiedz, jaka ktokolwiek dekodowal, byla ta dzialajaca przez pomylke.
Zabetonowany algorytm ukrywal nieprzetestowana sciezke.

Klucze slownika to teraz caly tag, wiec 7F49 jest 0x7F49. Testy sprawdzaja obie
odpowiedzi na ksztaltach, jakie karta faktycznie wysyla.

291 testow przechodzi.

+++ tests/Blinky.UnitTests/TlvTests.cs
@@ -52,5 +52,5 @@ public sealed class TlvTests
     private static Dictionary<byte, byte[]> ParseSimple(string hex) =>
         Tlv.ParseSimple(Convert.FromHexString(hex));
 
-    private static Dictionary<byte, byte[]> ParseBer(byte[] data) => Tlv.ParseBer(data);
+    private static Dictionary<int, byte[]> ParseBer(byte[] data) => Tlv.ParseBer(data);
 }
